fix: move lualine config to Lazy
